requirements.txt
# Core ML and Image Processing
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.5.0
numpy>=1.24.0

# Hugging Face Integration
accelerate>=0.20.0
datasets>=2.12.0

# Translation Services
deep-translator>=1.11.4
googletrans==4.0.0rc1

# Web Interface
gradio>=3.35.0
streamlit>=1.24.0

# Evaluation Metrics
nltk>=3.8.1
sacrebleu>=2.3.1

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Development and Testing
pytest>=7.4.0
black>=23.3.0
flake8>=6.0.0